+++ example/tree/lib/r.dart
@@ -10,10 +10,19 @@ class R {
   static const String mainImage = 'assets/main_image.jpg';
 
   static const String response = 'assets/response.json';
+
+  static const Map<String, String> $files = {
+    'mainImage': 'assets/main_image.jpg',
+    'response': 'assets/response.json',
+  };
 }
 
 class _Images {
   const _Images._();
 
   final String shieldHero = 'assets/images/shield_hero.jpg';
+
+  static const Map<String, String> $files = {
+    'shieldHero': 'assets/images/shield_hero.jpg'
+  };
 }