+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.1.0
+- feat: Can exclude files and folders from generation, pass the `exclude` option
+- feat: Can generate map files, pass the `create_map_files` option to `true`
+
 ## 1.0.0
 - feat: stable release
 - fix: typo in README.md